app/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.2
gunicorn==19.7.1
psycopg2==2.7.3.1
google-cloud-storage==1.6.0
scipy==1.0.0b1
pillow==4.2.1
tensorflow==1.2
pyyaml==3.12
google-api-python-client==1.6.4
google-cloud-pubsub==0.30.1
